MemberPoint 구현

+++ src/main/java/com/nhnacademy/book/point/domain/PointCondition.java
@@ -7,6 +7,8 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
+
 @Entity
 @Getter
 @Setter
@@ -25,9 +27,11 @@ public class PointCondition {
     @Column(nullable = false)
     private PointConditionName name;
 
+    // 회원가입시 5000p, 리뷰: 200p, 리뷰와 사진: 500p
     @Column
     private Integer conditionPoint;
 
+    // 도서구매시 기본 포인트 비율 5%
     @Column
-    private Integer conditionPercentage;
+    private BigDecimal conditionPercentage;
 }
